+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-
+### Fixed
+- setOption('label') in ControlGroupBuilder modified form instead of group
 
 ## [2.3.0] - 2023-05-16
 ### Added